Move downscaling_method_as_list to utils
